+++ web/src/pages/ToDo/ToDoTable.jsx
@@ -17,8 +17,7 @@ import { errorToString, utcStringToLocalDate } from "../../utils/func";
 
 import { ToDoTableRow } from "./ToDoTableRow";
 
-export function ToDoTable({ myTasks, pteamIds, cveIds }) {
-  const [page, setPage] = useState(0);
+export function ToDoTable({ myTasks, pteamIds, cveIds, page, setPage }) {
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const skip = useSkipUntilAuthUserIsReady();
 
@@ -129,4 +128,6 @@ ToDoTable.propTypes = {
   myTasks: PropTypes.bool.isRequired,
   pteamIds: PropTypes.arrayOf(PropTypes.string).isRequired,
   cveIds: PropTypes.arrayOf(PropTypes.string),
+  page: PropTypes.number.isRequired,
+  setPage: PropTypes.func.isRequired,
 };
